1. Added image title inside gallery 2. More css refactoring
