+++ src/App.js
@@ -6,8 +6,8 @@ import './App.css';
 import { BrowserRouter, Route, Switch } from 'react-router-dom';
 
 import Home from './Home';
-import Haiyan from './Haiyan';
-import Shira from './Shira';
+import Haiyan from './Component/Haiyan';
+import Shira from './Component/Shira';
 
 /*
         <header className="App-header">
