+++ src/Component/GalleryModal.css
@@ -1,31 +1,32 @@
 #image-container {
-  margin: .5vw;
+  margin: .5vw auto;
   display: flex;
   flex-wrap: wrap;
   justify-content: center;
-  background-color: #aab1111;
 }
 
 #image-container div {
   width: 300px;
   margin: 1vw;
-  display: flex;
   align-items: center;
   background-color: #714231;
-  justify-content: inherit;
-  padding: 1vw;
 }
 
 #image-container div img {
-  width: 95%;
+  margin-block-start: 10px;
+  width: 280px;
   height: auto;
   transition: 1s;
-  padding: 15px;
 }
 
 #image-container div img:hover {
   filter: grayscale(100%);
   transform: scale(1.1);
+}
+
+#image-container div p {
+  font-size: 1.5rem;
+  font-weight: 500;
 }
