+++ src/Component/Shira.js
@@ -1,38 +1,8 @@
 import React from 'react';
-import GalleryModal from './Component/GalleryModal';
-import shiraicon from './assets/shira/shira-icon.png';
+import GalleryModal from './GalleryModal';
+import Top from './Top';
+import shiraicon from '../assets/shira/shira-icon.png';
 import './Shira.css';
-
-const navitems = [
-  	{name: 'Smile',
-  	 page: 'Haiyan Page',
-    },
-  	{name: 'Food',
-  	 page: 'Shira Page',
-    },
-    {
-      name: 'Drawings',
-      page: 'Travel page',
-    },
-    {
-      name: 'Coming up',
-      page: 'Feature Page',  
-    }
-  ];
-
-/*
-        <nav>    
-          <ul id="topitem">
-                {
-                  navitems.map((interest, index) => (
-                    <li className="person" key={index}>
-                      <p>{interest.name}</p>
-                    </li>
-                  ))
-                }
-          </ul>
-        </nav>
-*/
 
 class Shira extends React.Component {
 
@@ -56,28 +26,40 @@ class Shira extends React.Component {
         {src:"flyingXiannv.jpg", alt:"No Title"},
         {src:"garageman.jpg", alt:"Car in Garage"},
         {src:"ourhouse.jpg", alt:"Our House: Kitchen, Bedrooms and Garage"}
-      ]
+      ],
+      navitems: [
+        {name: 'Smile',
+         page: 'Haiyan Page',
+        },
+        {name: 'Food',
+         page: 'Shira Page',
+        },
+        {
+          name: 'Drawings',
+          page: 'Travel page',
+        },
+        {
+          name: 'Coming up',
+          page: 'Feature Page',  
+        }
+       ],
+      iconSrc: {src: shiraicon,
+                alt: "Shira",
+                pname: "Shira"}
     };
   }
 
   render() {
   	return (
-      <div className="person-container">
-        <div className="person-icon-title-container">
-          <div id="person-icon-container">
-            <img src={shiraicon} className="person-img" alt="Shira"/>
-          </div>
-          <div className="person-title">
-             <p>Shira's World</p>
-          </div>
-        </div>
-
+      <div className="person-container">        
+        <Top topicon={this.state.iconSrc} navitems={this.state.navitems} />
         <section id="main">
-          <GalleryModal images={this.state.drawings}/>
+          <GalleryModal images={this.state.drawings} />
         </section>
   	  </div>);
   }
 }
 
 export default Shira;
+
